[#98988452] Provides help menu; Refactors; Adds `init`
* Provides help menu
  * `-o` opens online help
  * Help menu describes radiian
  * Help menu explains `init` command
  * `radiian init` fulfills [#99645328](https://www.pivotaltracker.com/story/show/99645328)
* Refactors directories
  * Removes `src/` and `build/`
  * Moves most files to `lib/`
  * Fulfills [#100058434](https://www.pivotaltracker.com/story/show/100058434)
* Improves package.json
  * Replaces "authors" with "contributors"
  * Adds "main", "files", "engines", and "preferGlobal"
  * Points "bin" & "main" to `radiian.js`
* Refactors files
  * **Please Note**: much of the renaming was forced by Commander.js
    in order to get it to work.
  * Renames `cli.js` to `radiian.js`
    * Replaces `cli.js` with `radiian.js` in README
    * `radiian.js` is now, for the most part, a help menu
  * Extracts UI into `radiian-init.js`
    * `radiian init` tells Commander to run `radiian-init`
  * Adjusts paths in `spec/` so that gulp will pass
  * Adjusts paths in Generator.js
