+++ lib/Generator.js
@@ -31,7 +31,7 @@
        */
       function getReader(template) {
         return function() {
-          return read(__dirname + '/../templates/' + template, 'utf8');
+          return read(__dirname + '/templates/' + template, 'utf8');
         };
       }
