+++ lib/questions.js
