+++ lib/radiian-init.js
@@ -1,23 +1,15 @@
-#!/usr/bin/env node
 (function() {
   'use strict';
 
-  var program = require('commander');
   var inquirer = require('inquirer');
-
-  var packageInfo = require('../package.json');
-  var generator = require('./lib/Generator');
-  var questions = require('./lib/questions');
-
-  program
-    .version(packageInfo.version)
-    .parse(process.argv);
+  var generator = require('./Generator');
+  var questions = require('./questions');
 
   var userHappy = false;
   var submittedAnswers = null;
 
   function clear() {
-    process.stdout.write("\u001b[2J\u001b[0;0H");
+    process.stdout.write('\u001b[2J\u001b[0;0H');
   }
 
   function go(questions) {
@@ -29,7 +21,7 @@
         message: function() {
           clear();
           return 'Please confirm that your answers are correct.\n' +
-              JSON.stringify(answers, null, '  ') +
+            JSON.stringify(answers, null, '  ') +
             '\nAre these answers correct?';
         },
         default: true
@@ -53,14 +45,15 @@
       });
     });
   }
+
   go(questions);
 
-  require('deasync').loopWhile(function(){
+  require('deasync').loopWhile(function() {
     return !userHappy;
   });
 
-  console.log("Generating playbook for an app named " + submittedAnswers.appName + " in ./ansible...");
+  console.log('Generating playbook for an app named ' + submittedAnswers.appName + ' in ./ansible...');
   generator.generate(submittedAnswers);
 
-  console.log("... done! Check out ./ansible, it's full of creamy goodness");
+  console.log('... done! Check out ./ansible, it\'s full of creamy goodness');
 }());
