+++ lib/regexes.js
