+++ lib/templates/ansible.cfg.mustache
