+++ lib/templates/aws_keys.mustache
