+++ lib/templates/destroy-old-nodes.yaml.mustache
