+++ lib/templates/ec2.ini.mustache
