+++ lib/templates/ec2.py
