+++ lib/templates/immutable.yaml.mustache
