+++ lib/templates/provision.sh.mustache
