+++ lib/templates/tag-old-nodes.yaml.mustache
