+++ package.json
@@ -1,8 +1,9 @@
 {
   "name": "radiian",
   "version": "0.1.1",
-  "description": "radiian generates an Ansible playbook for immutable infrastructure with AWS.",
-  "main": "lib/cli.js",
+  "description": "Radiian generates an Ansible playbook for immutable infrastructure with AWS.",
+  "bin": "lib/radiian.js",
+  "main": "lib/radiian.js",
   "scripts": {
     "test": "gulp test"
   },
@@ -13,16 +14,30 @@
   "keywords": [
     "ansible",
     "AWS",
-    "immutable infrastructure",
+    "immutable",
     "infrastructure",
     "devops",
-    "server management",
+    "server",
     "provisioning"
   ],
-  "author": "John F. Mercer <john@radify.io> and Gavin Davies <gavin@radify.io> (http://radify.io)",
+  "contributors": [{
+      "name": "John F. Mercer",
+      "email": "john@radify.io",
+      "url": "http://radify.io"
+    }, {
+      "name": "Gavin Davies",
+      "email": "gavin@radify.io",
+      "url": "http://radify.io"
+  }],
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/radify/radiian/issues"
+  },
+  "files": [
+    "lib"
+  ],
+  "engines": {
+    "node": ">=0.10.38"
   },
   "homepage": "https://github.com/radify/radiian#readme",
   "devDependencies": {
@@ -39,6 +54,8 @@
     "inquirer": "^0.9.0",
     "mkdirp": "^0.5.1",
     "mustache": "^2.1.2",
+    "openurl": "^1.1.0",
     "q": "^1.4.1"
-  }
+  },
+  "preferGlobal": true
 }
